pyproject.toml
[project]
name = "mcp_polygon"
version = "1.0.0"
description = "A MCP server project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.15.0",
    "polygon-api-client>=1.15.4",
    "websockets>=13.0",
]
[[project.authors]]
name = "Polygon"
email = "support@polygon.io"

[[project.maintainers]]
name = "Chris Scragg"
email = "clscragg@protonmail.com"

[project.urls]
Homepage = "https://github.com/ChrisSc/mcp_polygon"
Repository = "https://github.com/ChrisSc/mcp_polygon"
"Original Repository" = "https://github.com/polygon-io/mcp_polygon"
Documentation = "https://github.com/ChrisSc/mcp_polygon#readme"
Issues = "https://github.com/ChrisSc/mcp_polygon/issues"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.13.2",
    "pytest>=8.4.0",
    "pytest-asyncio>=0.24.0",
    "rust-just>=1.42.4",
]

[project.scripts]
mcp_polygon = "mcp_polygon:run"
